fix for triton2.2.0

+++ lmdeploy/pytorch/kernels/fused_moe.py
@@ -269,7 +269,7 @@ def _start_end_kernel(TopkIdx, SortedIdx, ExpStart, ExpEnd,
         sidx = tl.load(SortedIdx + sidx_off, mask=sidx_mask, other=0)
         tidx = tl.load(TopkIdx + sidx, mask=sidx_mask, other=num_experts)
         tidx_mask = tidx == exp_id
-        cnt += tl.sum(tidx_mask)
+        cnt += tl.sum(tidx_mask.to(tl.int32))
         if cnt > 0 and exp_start < 0:
             exp_start = sidx_start + tl.argmax(tidx_mask, axis=0)
